Check existance of package-lock.json before deleting to stop file not found spam in npm-utils.eclass.

+++ eclass/npm-utils.eclass
@@ -33,8 +33,7 @@ npm_audit_package_lock_update() {
 	local dir="${1}"
 	einfo "dir=${dir}"
 	pushd "${dir}" || die
-	#npm audit fix --force
-	rm package-lock.json
+	[ -e package-lock.json ] && rm package-lock.json
 	npm i --package-lock-only
 	popd
 }
